quadcopter: Update the throttle valid trigger
Increase the minimum value that is considered the trigger for making the
drone run its motors. This value happens to be above the value that is
retrieved when the RC signal is lost, which means the drone will be
consider a loss of signal as if it wasn't receiving a valid throttle.

Signed-off-by: Dede Dodu <sebnat94@hotmail.com>