drive_motors: Test driving the motors
Two new programs introduced for trying to drive the motors through the
ESCs. The first program is using the Servo library which is meant for
some basic testing. The second program is more advanced since it
controls the arduino ports directly. This gives much higher control in
driving the PWM that is driven through every ESC, allowing for a higher
frequency than the default 50Hz (20ms period).

Signed-off-by: Dede Dodu <sebnat94@hotmail.com>