quadcopter: Introducing main program
This is introducing the quadcopter program that will run on the drone in
order to control it.

Mainly, it's job is to read drone angles and rotational speeds from the
BNO055 sensor, read the desired angle from the RC controller, and
compute the PID output based on the error. Once the PID output is
computed, the corresponding pulse values are driven through the ESCs to
update motors' speed.

This global loop runs at 100Hz (10ms period) since the bottleneck is
getting IMU's data.

Signed-off-by: Dede Dodu <sebnat94@hotmail.com>